@@ -1,16 +1,12 @@
 package com.swmansion.rnscreens;
 
-import android.arch.lifecycle.Lifecycle;
-import android.arch.lifecycle.LifecycleObserver;
-import android.support.v4.app.Fragment;
-import android.util.Log;
+import androidx.lifecycle.Lifecycle;
+import androidx.lifecycle.LifecycleObserver;
+import androidx.fragment.app.Fragment;
+
 import android.view.View;
 import android.view.ViewParent;
 
-import com.facebook.react.modules.core.ChoreographerCompat;
-import com.facebook.react.modules.core.ReactChoreographer;
-
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
